% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullWorkshops.R
\name{installWorkshops}
\alias{installWorkshops}
\title{Install workshops from locations as indicated by main repository}
\usage{
installWorkshops(repository, location = "github", branch = "master")
}
\arguments{
\item{repository}{A single string indicating the 'username/repository' of the
GitHub repository}

\item{location}{Website location of the main repository (default 'github')}

\item{branch}{Indicates the branch from where to pull the DESCRIPTION file
(default 'master')}
}
\description{
This function looks at the 'issues' page of the main repository and installs
from the workshop URLs tagged as `[Workshop]`
}
