% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installWorkshops.R
\name{installWorkshops}
\alias{installWorkshops}
\title{Install workshops from locations as indicated by main repository}
\usage{

  installWorkshops(repository = workshopbuilder:::.options$get("BOOK_REPO"),
  local = workshopbuilder:::.options$get("REPOS_PATH"),
  location_url = "https://api.github.com/repos",
  ncpus = getOption("Ncpus", 1L), ...)
}
\arguments{
\item{repository}{A single string indicating the 'username/repository' of the
main book-building GitHub repository (defaults to package option
'BOOK_REPO')}

\item{location}{Website location of the main repository (default 'github')}
}
\description{
This function looks at the 'issues' page of the main repository and installs
from the workshop URLs tagged as `[Workshop]`
}
