% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installWorkshops.R
\name{installWorkshops}
\alias{installWorkshops}
\title{Install workshops from locations as indicated by main repository}
\usage{
installWorkshops(
  workshops = NULL,
  exclude = FALSE,
  repository = workshopbuilder:::.options$get("BOOK_REPO"),
  local = workshopbuilder:::.options$get("REPOS_PATH"),
  buildDir = "buildout",
  ncpus = 1L,
  ...
)
}
\arguments{
\item{workshops}{character() An optional vector of workshop repositories to
install. Default behavior installs all workshops.}

\item{exclude}{logical(1) Whether to include or exclude the workshop(s)
listed in the argument from the full list of workshops (default FALSE)}

\item{repository}{A single string indicating the 'username/repository' of the
main book-building GitHub repository (defaults to package option
'BOOK_REPO')}

\item{local}{character(1) The folder location for all cloned workshop
repositories}

\item{buildDir}{character(1) The folder location for all workshop
installation reports (within the 'local' folder)}

\item{ncpus}{integer(1) The number of CPUs to use for the building and
installation of workshop repositories (default: 1L)}
}
\description{
This function looks at the 'issues' page of the main repository and installs
from the workshop URLs tagged as `[Workshop]`
}
